Create BasesLocalesList component
