+++ pages/all.js
@@ -1,34 +1,16 @@
 import React, {useCallback} from 'react'
 import PropTypes from 'prop-types'
 import Router from 'next/router'
-import {Pane, Heading, Table, Paragraph, Button} from 'evergreen-ui'
+import {Pane, Heading, Paragraph, Button} from 'evergreen-ui'
 
 import {listBasesLocales} from '../lib/bal-api'
 
-import useFuse from '../hooks/fuse'
+import BasesLocalesList from '../components/bases-locales-list'
 
 function All({basesLocales}) {
-  const onBalSelect = useCallback(bal => {
-    if (bal.communes.length === 1) {
-      Router.push(
-        `/bal/commune?balId=${bal._id}&codeCommune=${bal.communes[0]}`,
-        `/bal/${bal._id}/communes/${bal.communes[0]}`
-      )
-    } else {
-      Router.push(`/bal?balId=${bal._id}`, `/bal/${bal._id}`)
-    }
-  }, [])
-
   const onCreate = useCallback(() => {
     Router.push('/new')
   }, [])
-
-  const [filtered, onFilter] = useFuse(basesLocales, 200, {
-    keys: [
-      'nom',
-      'commune'
-    ]
-  })
 
   return (
     <>
@@ -38,33 +20,8 @@ function All({basesLocales}) {
           Sélectionnez une Base Adresse Locale que vous souhaitez visualiser, créer ou modifier.
         </Paragraph>
       </Pane>
-      <Pane borderTop flex={1} overflowY='scroll'>
-        <Table>
-          <Table.Head>
-            <Table.SearchHeaderCell
-              placeholder='Rechercher une Base Adresse Locale'
-              onChange={onFilter}
-            />
-          </Table.Head>
-          {filtered.length === 0 && (
-            <Table.Row>
-              <Table.TextCell color='muted' fontStyle='italic'>
-                Aucun résultat
-              </Table.TextCell>
-            </Table.Row>
-          )}
-          <Table.Body background='tint1'>
-            {filtered.map(bal => (
-              <Table.Row key={bal._id} isSelectable onSelect={() => onBalSelect(bal)}>
-                <Table.TextCell>{bal.nom}</Table.TextCell>
-                <Table.TextCell flex='0 1 1'>
-                  {bal.communes.length < 2 ? `${bal.communes.length} commune` : `${bal.communes.length} communes`}
-                </Table.TextCell>
-              </Table.Row>
-            ))}
-          </Table.Body>
-        </Table>
-      </Pane>
+
+      <BasesLocalesList basesLocales={basesLocales} />
 
       <Pane borderTop marginTop='auto' padding={16}>
         <Paragraph size={300} color='muted'>
